pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ecommerce-platform"
version = "1.0.0"
description = "基于Flask + MySQL + Redis的完整电商平台"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Framework :: Flask",
]
dependencies = [
    "Flask==3.0.0",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Werkzeug==3.0.1",
    "redis==5.0.1",
    "PyMySQL[rsa]",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-flask",
    "black",
    "flake8",
    "mypy",
]
test = [
    "pytest>=7.0",
    "pytest-flask",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/yourusername/ecommerce-platform"
Repository = "https://github.com/yourusername/ecommerce-platform"
Issues = "https://github.com/yourusername/ecommerce-platform/issues"

[project.scripts]
ecommerce = "run:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true